+++ src/enums/metaverses.enum.ts
@@ -3,10 +3,3 @@ export enum Metaverses {
   Decentraland = 'decentraland',
   SomniumSpace = 'somnium-space',
 }
-
-export const METAVERSE_LABEL: Record<Metaverses | "All", string> = {
-  All: "All Lands",
-  [Metaverses.SandBox]: "The Sandbox",
-  [Metaverses.Decentraland]: "Decentraland",
-  [Metaverses.SomniumSpace]: "Somnium Space"
-} as const;
